refactor: clean up chain config files
